CyclicBarrierMain 例子
